move geometry out of rig as just xy math
