+++ pacman/operations/rig_algorithms/geometry.py
